Update DB for voting preference lists
SQL commands to update the votes TABLE to hold multiple votes per
election-person tuple.  This is achieved with a new column votes.id,
with an associated sequence such that we can store a list 'A C D B E'
and preserve the order when querying the votes per person per election
for use in a 'borda count' or similar 'specify a preference list' voting
system.