Display multiple votes per voter correctly
This code encompasses necessary changes to have the 'Election results' web
page show votes per voter in a nice format.

NB: Although in a single test the multiple votes for a voter came out in
the correct preference order it's not clear if that was chance or will
always work.